+++ R/family.R
@@ -81,15 +81,21 @@ epidist_family_param <- function(family, ...) {
 #' @importFrom cli cli_abort
 #' @export
 epidist_family_param.default <- function(family, ...) {
-  df <- data.frame(y = c(1, 2))
-  dummy_mdl <- make_stancode(y ~ 1, data = df, family = class(family)[1])
+  data_dummy <- data.frame(y = c(1, 2))
+  dummy_mdl <- make_stancode(
+    y ~ 1,
+    data = data_dummy,
+    family = class(family)[1]
+  )
 
   # get the lowered family name
   family_name <- tolower(class(family)[1])
 
   # Extract the Stan parameterisation from the dummy model code
   lpdf_pattern <- paste0(
-    "target \\+= ", family_name, "_(lpdf|lpmf)\\(Y \\| ([^)]+)\\)" # nolint
+    "target \\+= ", # nolint
+    family_name,
+    "_(lpdf|lpmf)\\(Y \\| (.+?)\\)" # nolint
   )
   lpdf_match <- regexpr(lpdf_pattern, dummy_mdl)
   reparam <- if (lpdf_match > 0) {
@@ -103,8 +109,11 @@ epidist_family_param.default <- function(family, ...) {
     match_str <- mu_matches[1]
     param <- sub(
       paste0(
-        "target \\+= ", family_name, "_(lpdf|lpmf)\\(Y \\| " # nolint
-      ), "",
+        "target \\+= ", # nolint
+        family_name,
+        "_(lpdf|lpmf)\\(Y \\| " # nolint
+      ),
+      "",
       match_str
     )
     param <- sub(")", "", param, fixed = TRUE)