+++ R/globals.R
@@ -5,7 +5,5 @@ utils::globalVariables(c(
   "woverlap", # <epidist_stancode.epidist_latent_model>
   "rlnorm", # <simulate_secondary>
   "fix", # <.replace_prior>
-  "prior_new", # <.replace_prior>
-  "source_new", # <.replace_prior>
   NULL
 ))